+++ Assets/Prefabs/Entity/Enemy.prefab
@@ -907,10 +907,9 @@ GameObject:
   m_Component:
   - component: {fileID: 823462048605159767}
   - component: {fileID: 2774912599851552801}
-  - component: {fileID: 6003420987359868496}
-  - component: {fileID: 3008665686695201217}
-  - component: {fileID: 1126781104185213411}
   - component: {fileID: 1195520281710921038}
+  - component: {fileID: 4248805295292920887}
+  - component: {fileID: 6518899734363425440}
   m_Layer: 0
   m_Name: Enemy
   m_TagString: Untagged
@@ -947,68 +946,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 53ab215f512bcf240ad1f6ee7b19c927, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  ResourceMan: {fileID: 0}
+  StatMan: {fileID: 0}
   _healthGauge: {fileID: 111669627266848594}
   _actionGauge: {fileID: 3714951229989132436}
   CombatMode: 0
   CombatTimeout: 5
   Alive: 1
---- !u!114 &6003420987359868496
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6127087731749915755}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5d999dfcb898c8d43b4310178fe747c6, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  Value: 0
-  BaseValue: 20
-  MaxValue: 100
-  TotalValue: 0
-  RechargeInterval: 1
-  RechargeRate: 2
-  RechargeModifier: 2
---- !u!114 &3008665686695201217
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6127087731749915755}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2231e70241472364ba0822d750e0495f, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  Value: 0
-  BaseValue: 20
-  MaxValue: 100
-  TotalValue: 0
-  RechargeInterval: 1
-  RechargeRate: 2
-  RechargeModifier: 1
---- !u!114 &1126781104185213411
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6127087731749915755}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8d205df20b784014b82f34a924d42edd, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  Value: 0
-  BaseValue: 20
-  MaxValue: 100
-  TotalValue: 0
-  RechargeInterval: 1
-  RechargeRate: 2
-  RechargeModifier: 1
 --- !u!54 &1195520281710921038
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -1036,6 +980,31 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 0
   m_CollisionDetection: 0
+--- !u!114 &4248805295292920887
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6127087731749915755}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 79119190a5f7bf54e968846b69a279ef, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _initialData: {fileID: 11400000, guid: 48cb9076620a6f14abbc2ca193121550, type: 2}
+--- !u!114 &6518899734363425440
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6127087731749915755}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 29f960be507f84040ac93135a4b95296, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &6228925391245955044
 GameObject:
   m_ObjectHideFlags: 0
