+++ Assets/Scripts/Mechanic/EntityInfo/Resource/ResourceManager.cs
@@ -4,26 +4,37 @@ using UnityEngine;
 
 public class ResourceManager : MonoBehaviour
 {
-    private Dictionary<EntityDataTypes.Resource, ResourceTemplate> _resourceDict;
-    public Dictionary<EntityDataTypes.Resource, ResourceTemplate>  ResourceDict {  get { return _resourceDict; } }
+    /*private Dictionary<EntityDataTypes.Resource, ResourceTemplate> _resourceDict = new Dictionary<EntityDataTypes.Resource, ResourceTemplate>();
+    public Dictionary<EntityDataTypes.Resource, ResourceTemplate>  ResourceDict {  get { return _resourceDict; } }*/
 
-    public ResourceTemplate Health { get{ return _resourceDict[EntityDataTypes.Resource.HP]; } }
-    public ResourceTemplate Action { get{ return _resourceDict[EntityDataTypes.Resource.AP]; } }
+    public Health Health { get{ return _health; } }
+    public Action Action { get{ return _action; } }
+    public Currency Currency { get{ return _currency; } }
 
-    protected void Awake()
+    private Health _health;
+    private Action _action;
+    private Currency _currency;
+
+    public void Initialize()
     {
-        if (!_resourceDict.ContainsKey(EntityDataTypes.Resource.HP))
-            _resourceDict[EntityDataTypes.Resource.HP] = new Health();
-        if (!_resourceDict.ContainsKey(EntityDataTypes.Resource.AP))
-            _resourceDict[EntityDataTypes.Resource.AP] = new Action();
+        if (_health == null)
+            _health = new Health();
+        if (_action == null)
+            _action = new Action();
+    }
+
+    public void AddCurrency()
+    {
+        if (_currency == null)
+            _currency = new Currency();
     }
 
     public void UpdateResourceStats(float hp, float ap)
     {
-        _resourceDict[EntityDataTypes.Resource.HP].TotalValue = hp;
-        _resourceDict[EntityDataTypes.Resource.AP].TotalValue = ap;
+        _health.TotalValue = hp;
+        _action.TotalValue = ap;
     }
-
+/*
     public float GetValue(EntityDataTypes.Resource resource)
     {
         return _resourceDict[resource].Value;
@@ -32,5 +43,5 @@ public class ResourceManager : MonoBehaviour
     public float GetTotalValue(EntityDataTypes.Resource resource)
     {
         return _resourceDict[resource].TotalValue;
-    }
+    }*/
 }
