--- Assets/Scripts/Mechanic/EntityInfo/Stats/ScriptableObjects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ad22bca5d5e9c16499cdff3c167ba94a
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
