--- Assets/Scripts/Mechanic/EntityInfo/Stats/ScriptableObjects/EntityDataScriptableObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: efa0ef1f222681646923a01cec4103af
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
