--- Assets/Scripts/Mechanic/EntityInfo/Stats/ScriptableObjects/UpgradeLevelsScriptableObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 08a3e0b0f3f6542448552d35549813f8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
