--- Assets/Scripts/Mechanic/EntityInfo/Stats/ScriptableObjects/UpgradeScriptableObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7437c2ce5d76d6f40b78121816991b4e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
