+++ Assets/Scripts/Mechanic/Upgrades.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0aca1fc5d296bd48bd5408f71ac198d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
