+++ Assets/Scripts/SOJsonConverter - Copy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 94a30c7bd9fa7904c80d466cbf6d39cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
