+++ Assets/Scripts/SOJsonConverter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55cc611d293aaad49b057d5c1ed49d6b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
