+++ Assets/StatData/EnemyData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 48cb9076620a6f14abbc2ca193121550
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
