+++ Assets/StatData/PlayerData.asset
@@ -36,22 +36,22 @@ MonoBehaviour:
   _maxDefence: 1e+9
   _atk_speed: 10
   _minAtk_speed: 1
-  _maxAtk_speed: 0
+  _maxAtk_speed: 1e+9
   _accuracy: 100
   _minAccuracy: 0
-  _maxAccuracy: 0
+  _maxAccuracy: 1e+9
   _evasiveness: 0
   _minEvasiveness: 0
-  _maxEvasiveness: 0
+  _maxEvasiveness: 1e+9
   _crit_dmg: 200
   _minCrit_dmg: 0
-  _maxCrit_dmg: 0
+  _maxCrit_dmg: 1e+9
   _crit_rate: 5
   _minCrit_rate: 0
-  _maxCrit_rate: 0
+  _maxCrit_rate: 1e+9
   _speed: 1
   _minSpeed: 0
-  _maxSpeed: 0
+  _maxSpeed: 1e+9
   _luck: 0
   _minLuck: 0
   _maxLuck: 1e+9
