+++ Assets/StatData/RangedEnemy.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f16b430c7ee25f4d8ea23a53662a59a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
